+++ eurorack/src/lib.rs
@@ -1,5 +1,8 @@
 #![no_main]
 #![no_std]
+#![feature(alloc_error_handler)]
+
+extern crate alloc;
 
 use defmt_rtt as _; // global logger
 use panic_probe as _;
@@ -12,6 +15,11 @@ pub mod system;
 #[defmt::panic_handler]
 fn panic() -> ! {
     cortex_m::asm::udf()
+}
+
+#[alloc_error_handler]
+fn oom(_: core::alloc::Layout) -> ! {
+    panic!("OOM");
 }
 
 /// Terminates the application and makes `probe-run` exit with exit-code = 0